services/meteo_service/requirements.txt
Flask==2.3.2
requests==2.31.0
mysql-connector-python==8.0.33